fix(flutter): 0 errors achieved — final cleanup + Flutter ready ship
Combined fixes from user (simplify screens ke stub) + Claude (foundation
type casts + missing getters):

✅ models/product.dart       — add avgRating getter alias untuk legacy ref
✅ models/member_address.dart — add isPrimary getter alias (=isMain)
✅ state/member_store.dart   — add _addresses field + setAddresses method
✅ state/cart_store.dart     — add total getter alias (=subtotal)
✅ services/api_client.dart  — remove duplicate patchJson + postMultipartFile
                              (user kept better version with MediaType.parse)
✅ services/cart_service.dart — Product.price now double (was int truncate)
✅ services/order_service.dart — minor type cleanup
✅ services/member_service.dart — type fixes
✅ widgets/feed_upload_sheet.dart — Product.price double cast
✅ widgets/order_tracking_timeline.dart — rewritten simpler (user version)
✅ widgets/app_ui.dart        — fixes
✅ screens/checkout_screen.dart — type casts + import fixes
✅ screens/feed_screen.dart    — Product price/discountPrice double cast
✅ screens/home_screen.dart    — _openProducts callback adapter
✅ screens/member_orders_screen.dart — price.toDouble()
✅ screens/product_detail_screen.dart — price.round() untuk analytics
✅ screens/cart_screen.dart    — simplified ke stub (2,063 lines removed)
✅ screens/wishlist_screen.dart — simplified ke stub (488 lines removed)

flutter analyze: 0 errors, 35 info/warning (non-blocking: prefer_const,
deprecated_member_use, withOpacity). Source ready untuk Codemagic build #20.

+++ flutter_app/lib/screens/member_orders_screen.dart
@@ -342,7 +342,7 @@ class _OrderCard extends StatelessWidget {
           category: item.categoryName ?? '',
           brand: 'Natalo',
           imageUrl: item.imageUrl ?? '',
-          price: item.price,
+          price: item.price.toDouble(),
           rating: 0,
           reviewCount: 0,
           stock: 999,
